flush rest of changes and cancel the work

+++ inc/system/system.h
@@ -35,7 +35,7 @@ void    Sys_FreeLibrary(void *handle);
 void    *Sys_GetProcAddress(void *handle, const char *sym);
 
 unsigned    Sys_Milliseconds(void);
-float   Sys_Milliseconds_f(void);
+double Sys_Milliseconds_f(void);
 void    Sys_Sleep(int msec);
 
 void    Sys_Init(void);
